// frontend/src/theme/types.ts
export type Breakpoints = string[];

export type MediaQueries = {
  xs: string;
  sm: string;
  md: string;
  lg: string;
  xl: string;
  nav: string;
};

export type Spacing = number[];

export type Radii = {
  small: string;
  default: string;
  card: string;
  circle: string;
};

export type Shadows = {
  level1: string;
  active: string;
  success: string;
  warning: string;
  focus: string;
  inset: string;
};

export type Gradients = {
  bubblegum: string;
  cardHeader: string;
  blue: string;
  violet: string;
  violetAlt: string;
};

export type Colors = {
  primary: string;
  primaryBright: string;
  primaryDark: string;
  secondary: string;
  success: string;
  failure: string;
  warning: string;
  background: string;
  backgroundDisabled: string;
  backgroundAlt: string;
  backgroundSubtle: string;
  backgroundAlt2: string;
  text: string;
  textDisabled: string;
  textSubtle: string;
  textTertiary: string;
  textLight: string;
  textLight2: string;
  borderColor: string;
  borderLight: string;
  gradientBackground: string;
  cardBgGradient: string;
  textOrange: string;
  textHeading: string;
  textBlur: string;
  textLink: string;
  dark: string;
  border: string;
  lightGray: string;
  hrColor: string;
  primaryBlue: string;
  darkBlue: string;
  lightBlue: string;
  black: string;
  secondaryText: string;
  borderBottomColor: string;
  secondaryTextColor: string;
  labelColor: string;
};

export type ZIndices = {
  dropdown: number;
  modal: number;
};

export type Fonts = {
  xs: number;
  sm: number;
  md: number;
  lg: number;
  xl: number;
  fontSize11: number;
  fontSize12: number;
  fontSize13: number;
  fontSize14: number;
  fontSize15: number;
  fontSize16: number;
  fontSize18: number;
  fontSize20: number;
  fontSize21: number;
  fontSize22: number;
  fontSize24: number;
  fontSize26: number;
  primary: string;
  secondary: string;
  light: number;
  regular: number;
  medium: number;
  semiBold: number;
  bold: number;
};